Version 1.0.0: September 3, 2008
Added nearest-neighbor interchanges (NNIs) according to a
minimum-evolution criterion. This is done in a similar way as in
FastME, but FastTree uses profiles to speed the computation and
uses weighted joins (inspired by BIONJ) to compute profiles of
internal nodes.  This change gives significant improvements in the
accuracy of FastTree, partly because the NNIs are beneficial and
partly because FastTree corrects distances for multiple
substitutions (e.g., uses Jukes-Cantor or Poisson-corrected
distances) when doing NNIs but not during the initial
neighbor-joining phase.

Optimized the implementation of local bootstrap and replaced the
probabilistic local supports, which do not transfer well to
log-corrected distances, with the local bootstrap.

Eliminated crashes on alignments with large numbers of
highly-gapped and non-overlapping sequences, by adding checks to
avoid divide-by-zero errors.