Version 1.1.0: May 11, 2009
Added heuristic subtree-prune-regraft moves. As suggested by
Richard Desper and Olivier Gascuel, each SPR is treated as a chain
of NNIs, and FastTree uses the sum of changes in lengths of
internal branches as an estimate of the change in tree length for
the SPR. (In FastME's balanced NNIs, the change in internal branch
length is the same as the change in tree length, but for FastTree,
this is an approximation because of the log-of-averages
vs. averages-of-logs issue.) By default, FastTree does two rounds
of SPRs, with exhaustive search for moves up to length two. Then
it extends each of these moves up to a length of 10 along the best
step at each point. (You can use -spr 0 to turn these off, as they
are a bit slow and only seem to improve accuracy slightly.)

Doubled the default number of rounds of NNIs.

Switch to using balanced joins by default, as they are much faster
and have no apparent impact on accuracy.

Tweaked the top-hits heuristic to be more aggressive -- the
default -close value is now log(N)/(log(N)+2). This makes it
noticeably faster on huge alignments, without any impact on
accuracy.

Removed O(N**2) steps -- use a sqrt(N) "top-visible" list to
select a candidate for the best join in O(sqrt(N)) time instead of
O(N) time per join. Also update the total profile from scratch much
less frequently.

Use stale out-distances more aggressively, and rescale them by the
number of active nodes. (The intuition behind this optimization is
that the average out-distance changes very little after a join if
we have 1,000 active nodes.)

Optimized the constrained topology search to be faster. The
constraintWeight is now rather ad hoc, but in practice, a weight
of 10 means the constraints are always satisfied.

Warn if appear to be analyzing a protein alignment with -nt or a
nucleotide alignment without -nt

More compact help

Compact reporting of unexpected characters

Progress indicator (unless running with -quiet)