Version 1.9.0: June 19, 2009
Added maximum-likelihood nearest-neighbor interchanges, with the
Jukes-Cantor model for nucleotide data or the Jones-Taylor-Thorton
model for amino acid data, and no variation in rates across sites.
By default, do O(log N) rounds of maximum-likelihood NNIs, but
quit if the total tree log-likelihood has converged (it improves
by less than 0.1 over the last round).

When considering NNI moves (3 possible quartets), FastTree
optimizes all 5 branch lengths in turn, up to a branch length
accuracy of 0.001. By default it uses two rounds of optimizing the
5 branch lengths to estimate the log-likelihood for the
quartet. During the first round of NNIs, the minimum-evolution
profiles and branch lengths are used initially, but they are
gradually replaced by the posterior distributions and the
maximum-likelihood branch lengths.

Two heuristics to speed up ML NNIs: the star topology test and the
one-round heuristic.

Star topology test: If ((A,B),(C,D)) is noticeably more likely
than the star topology (the log-likelihood with the optimal
internal branch length is at least 5 better), and there was no NNI
at this node or its neighbors in the previous round, then do not
optimize the other branch lengths or estimate log likelihoods for
`the other topologies.

One-round heuristic: If after the 1st round of optimizing 5 branch
lengths for the 3 possible quartet topologies, the log-likelihood
value is more than 5 worse than the original topology, do not
optimize that topology further (no 2nd round of optimizing branch
lengths).

For amino acid data, uses an eigen-representation of posterior
distributions at internal nodes to compute likelihoods in O(a)
time, not O(a***2) time, where a is the alphabet size. However,
computing posterior distributions still takes O(a**2) time. Uses
an approximation to reduce this to O(a) time if the posterior
distribution is dominated by one amino acid and the approximation
is accurate.

For the Jukes-Cantor model, the posterior distribution of a
character and a gap, or of two characters that match (potentially
also mixed with gaps or uncertainty) is the same as another
character mixed with a gap. FastTree uses this representation to
save memory and CPU time.

Added many command-line options relating to maximum-likelihood
NNIs, such as -mlnni to change the maxinum number of rounds of ML
NNIs, -mlacc 2 to turn off the ML heuristics (or higher values to
optimize the branch lengths for a quartet of posterior profiles
more thoroughly), -mlexact to turn off approximate posteriors,
-noml to turn off ML computations entirely, and -nome to turn off
minimum-evolution computations entirely (except that
minimum-evolution is still used to get initial branch lengths).
change the number of rounds of ML NNIs.)

Make sure profiles are updated on the way "back up" during NNIs