Version 2.0.0: July 22, 2009
Accounted for variable rates across sites with the CAT
approximation to the gamma model used by RAxML (see A. Stamitakis,
"Phylogenetic models of rate heterogeneity: a high performance
computing perspective", IPDPS 2006). The site categories are set
once, after the first round of NNIs or (if using -mllen) after the
first round of optimizing branch lengths. FastTree uses a Bayesian
approach to avoid overfitting these rates.

Implemented SH-like local supports, as in PhyML 3. In practice,
FastTree's supports are nearly identical to those from PhyML with
the discrete Gamma-4 model, despite the CAT approximation. One
difference is that FastTree reports 0 for a "bad split" that turns
out to support a different topology, instead of reporting the
(negative) likelihood difference as the support value. During this
phase, FastTree also reports any "bad" splits (NNIs that would
improve the likelihood) and whether these are due to constraints
or not.

Implemented the generalized time-reversible nucleotide model of
evolution (-gtr option). The base frequencies are the same as in
the alignment; the rates are optimized just before selecting site
categories. Or use -gtrfreq and -gtrrates to set them yourself.

Improvements to heuristics for maximum-likelihood NNIs:
(1) Fixed a bug in the code for deciding whether to skip the 2nd
round of optimizing branch lengths around a quartet.
(2) Skip the 2nd round if one of the NNIs is a clear winner, not
just if the original topology is a clear winner.
(3) Skip the 2nd round if we are allowing the star test, the
original topology is winning, and the internal branch lengths of
the alternatives are roughly 0, as a biologically significant
improvement is unlikely.
(4) Skip traversing into an entire subtree if no significant
improvement in that subtree has been found in either of the last
two rounds. This affects minimum-evolution NNIs as well.  Use
-slownni to turn off subtree skipping.
(5) This allows us to increase the default number of rounds of
minimum-evolution NNIs to 4*log2(N) and the default number of
rounds of maximum-likelihood NNIs to 2*log2(N)..
(6) Stop doing NNIs if no significant tree improvements were found
in the last round.
(7) But, add a single slow round of ML NNIs at the end to catch
the rare errors that result from all the heuristics.

The -mllen option, in combination with -intree and -nome, lets you
optimize branch lengths for a fixed topology.

The -log option saves intermediate trees, settings, and model
details to the specified log file.

Use SSE3 instructions if the compiler has set __SSE__ to indicate
that they are available. SSE3 improves performance of protein
maximum likelihood up to 50% and speeds up the protein
minimum-evolution code slightly. Nucleotide code may be slightly
faster or slightly slower. Use -DNO_SSE to compile without SSE
instructions.

Sped up convergence in the branch length optimizer by tweaking the
selection of starting bounds for the brent method and by
terminating optimization if the new guess differs by less than
0.2% of the length or less than 0.0001.

Reduce posterior profile computations while optimizing a quartet
from 10 to 7 (AB, CD, ABC, ABD, CD with new branch lengths, ACD,
BCD).

Raised default -constraintWeight to 100 because 10 is not that big
a difference in log likelihoods.

Use exact posteriors for protein sequences by default.