Version 2.1.10: April 11, 2017
Fix a bug when using GTR models with huge alignments with over 2
billion non-gap characters. The frequencies of A, C, G, and T were
being computed with a 32-bit (signed) counter. This led to
negative frequencies and eventually to a crash with
"FastTree.c:9769: tqli: Assertion `iter < 30' failed.". SetMLGtr()
now uses 64-bit counters. Also, more information about the
optimization of the GTR model is saved in the log file (if using
the -log option). To support this, gtr_opt_t now includes fpLog.